Consolidate files in UI package
